+++ lib/body.js
@@ -7,7 +7,7 @@
 var convert = require('encoding').convert;
 var bodyStream = require('is-stream');
 var PassThrough = require('stream').PassThrough;
-var FetchError = require('./error');
+var FetchError = require('./fetch-error');
 
 module.exports = Body;
