+++ lib/fetch-error.js
@@ -1,5 +1,5 @@
 /**
- * error.js
+ * fetch-error.js
  *
  * Helper function for creating a FetchError
  */
